% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{convert}
\usage{
convert(df = NULL, hasFlanks = FALSE, addFlanks = FALSE,
  panel = "ForenSeq", format = "UAS", lusSep = "_")
}
\arguments{
\item{df}{A dataframe which must contain the following columns: Locus, Sequence (upper case Strings), other columns are kept}

\item{hasFlanks}{Whether flanks are part of sequence (should be true if format=STRaitRazor)}

\item{addFlanks}{Whether flanks should be added to bracket format}

\item{panel}{Assumed panel (ForenSeq). PowerPlex not yet implemented}

\item{format}{Format of data (UAS or STRaitRazor): Assumes that STRaitRazor (and others) are already in forward direction, while UAS must revComp for some markers}

\item{lusSep}{Separator for LUS format}
}
\value{
A dataframe with an additional column bracket format (converted sequences)
Traditional_STR_Allele: Traditional CE/RU format
Forward_Strand_Bracketed_form: Bracket format format in forward direction
UAS_Output_Bracketed_Form: Bracket format in forward/backward direction wrt UAS definitions
LUS_Plus: Provides LUS+ format of sequences
}
\description{
Convert sequences to bracket format
}
\details{
The dataframe must contain the following columns: Locus, Sequence, others are kept
This is the main function for converting sequences into bracket format as done by lusSTR
}
\examples{
\dontrun{ 
pgkPath <- path.package("LUSstrR") # Get package path.
df = readRDS(paste0(pgkPath,"/examples/exampleDataframe.RDS"))
df = convert(df)
}
}
\author{
Oyvind Bleka
}
