% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFromUASreport.R
\name{convertFromUASreport}
\alias{convertFromUASreport}
\title{convertFromUASreport}
\usage{
convertFromUASreport(file, AT = 11, typedOnly = FALSE,
  hasFlanks = FALSE)
}
\arguments{
\item{file}{File name of report}

\item{AT}{A list of allele frequencies for a given population.}

\item{typedOnly}{Whether only typed alleles should be converted}

\item{hasFlanks}{Whether flanks are part of sequence}
}
\value{
A dataframe with an additional column bracket format (converted sequences)
}
\description{
Function which converts sequences to bracket format for UAS report file input
}
\details{
This function is a data format wrapper
}
\examples{
\dontrun{ 
pgkPath <- path.package("LUSstrR", quiet = TRUE) # Get package path.
file = paste0(pgkPath,"/examples/UAS_Sample_Details_Report_test.xlsx")
df = convertFromUASreport(file)
}
}
\author{
Oyvind Bleka
}
